pyproject.toml
[project]
name = "cv-builder-from-yaml-to-pdf"
version = "0.1.0"
description = "A tool to generate PDF CVs from YAML files"
authors = [
    {name = "yantology",email = "work@yantology.dev"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "pyyaml>=6.0",
    "reportlab>=3.6.12",
    "click>=8.1.3",
    "jinja2>=3.1.2",
    "pydantic (>=2.11.4,<3.0.0)",
]

[tool.poetry]
packages = [{include = "cv_builder_from_yaml_to_pdf", from = "src"}]

[project.scripts]
cv-builder = "cv_builder_from_yaml_to_pdf.main:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
